refactor(core): move pipeline metrics out of lib
