+++ docs/topology-status.md
@@ -25,13 +25,13 @@ Generated by `devtools render-topology-status`. Reads `docs/plans/topology-targe
 | archive-semantic | — | 5 | 5 | 0 | 0 |
 | archive-session | — | 6 | 6 | 0 | 0 |
 | archive-viewport | — | 5 | 5 | 0 | 0 |
-| core-primitive | — | 8 | 8 | 0 | 0 |
+| core-primitive | — | 9 | 9 | 0 | 0 |
 
 ### Summary
 
 - **Stable** (no move scoped): 579
 - **Kernel** (polylogue/ root): 10
-- **Primitives** (lib-root + storage-root): 22
+- **Primitives** (lib-root + storage-root): 21
 - **TBD** (cell needs explicit assignment): 0
 - **Total declared**: 697
 - **Realized polylogue/**/*.py**: 697 files declared
