+++ polylogue/sources/source_acquisition_components.py
@@ -13,7 +13,7 @@ from polylogue.archive.artifact_taxonomy import classify_artifact
 from polylogue.config import Source
 from polylogue.core.json import JSONDocument, JSONValue, is_json_value
 from polylogue.core.json import dumps_bytes as json_dumps_bytes
-from polylogue.lib.metrics import read_current_rss_mb, read_peak_rss_self_mb
+from polylogue.core.metrics import read_current_rss_mb, read_peak_rss_self_mb
 from polylogue.storage.blob_store import BlobStore
 from polylogue.storage.cursor_state import CursorStatePayload
 from polylogue.types import Provider
